+++ photutils/aperture/rectangle.py
@@ -149,6 +149,8 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
         self.w = float(w)
         self.h = float(h)
         self.theta = float(theta)
+        self._repr_params = [('w', self.w), ('h', self.h),
+                             ('theta', self.theta)]
 
     @property
     def bounding_boxes(self):
@@ -250,6 +252,9 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
         self.h_out = float(h_out)
         self.h_in = self.w_in * self.h_out / self.w_out
         self.theta = float(theta)
+        self._repr_params = [('w_in', self.w_in), ('w_out', self.w_out),
+                             ('h_in', self.h_in), ('h_out', self.h_out),
+                             ('theta', self.theta)]
 
     @property
     def bounding_boxes(self):
@@ -343,6 +348,8 @@ class SkyRectangularAperture(SkyAperture):
         self.w = w
         self.h = h
         self.theta = theta
+        self._repr_params = [('w', self.w), ('h', self.h),
+                             ('theta', self.theta)]
 
     def to_pixel(self, wcs, mode='all'):
         """
@@ -443,6 +450,8 @@ class SkyRectangularAnnulus(SkyAperture):
         self.w_out = w_out
         self.h_out = h_out
         self.theta = theta
+        self._repr_params = [('w_in', self.w_in), ('w_out', self.w_out),
+                             ('h_out', self.h_out), ('theta', self.theta)]
 
     def to_pixel(self, wcs, mode='all'):
         """